+++ src/mpservice/mpserver.py
@@ -91,9 +91,15 @@ from time import perf_counter, sleep
 from typing import Sequence, Union, Callable, Type, Any
 
 import psutil
+from overrides import final
 
-from .util import RemoteException, exit_err_msg, is_exception
-from .util import Thread, SpawnProcess, TimeoutError
+from .util import (
+    MP_SPAWN_CTX, SpawnProcess,
+    Thread,
+    RemoteException,
+    exit_err_msg, is_exception,
+    TimeoutError,
+)
 from ._queues import SingleLane
 
 # This modules uses the 'spawn' method to create processes.
@@ -281,6 +287,7 @@ class Worker(metaclass=ABCMeta):
         self._batch_buffer = SingleLane(self.batch_size + 10)
         self._batch_get_called = threading.Event()
         collector_thread = Thread(target=self._build_input_batches, args=(q_in, q_out))
+        collector_thread.start()
 
         n_batches = 0
         batch_size_log_cadence = self.batch_size_log_cadence
@@ -528,6 +535,7 @@ class ProcessServlet:
                     **self._init_kwargs,
                 },
             ))
+            self._workers[-1].start()
             name = q_out.get()
             logger.debug(f"   ... worker <{name}> is ready")
 
@@ -570,6 +578,7 @@ class ThreadServlet:
                     **self._init_kwargs,
                 },
             )
+            w.start()
             self._workers.append(w)
             name = q_out.get()
             logger.debug(f"   ... worker <{name}> is ready")
@@ -664,8 +673,12 @@ class Ensemble:
                 s.start(q1, q2)
             self._qins.append(q1)
             self._qouts.append(q2)
-        self._threads.append(Thread(target=self._dequeue))
-        self._threads.append(Thread(target=self._enqueue))
+        t = Thread(target=self._dequeue)
+        t.start()
+        self._threads.append(t)
+        t = Thread(target=self._enqueue)
+        t.start()
+        self._threads.append(t)
         self._started = True
 
     def _enqueue(self):
@@ -799,8 +812,12 @@ class Server:
             self._q_out = FastQueue()
         self._servlet.start(self._q_in, self._q_out)
 
-        self._threads.append(Thread(target=self._gather_output))
-        self._threads.append(Thread(target=self._onboard_input))
+        t = Thread(target=self._gather_output)
+        t.start()
+        self._threads.append(t)
+        t = Thread(target=self._onboard_input)
+        t.start()
+        self._threads.append(t)
 
         self._started = True
         return self
@@ -823,14 +840,20 @@ class Server:
         self._started = False
 
     @classmethod
-    def get_mpcontext(cls):
+    @final
+    def get_mp_context(cls):
         # If subclasses need to use additional Queues, Locks, Conditions, etc,
         # they should create them out of this context.
         # Subclass should not customize this method. Always use spawned processes.
-        return multiprocessing.get_context('spawn')
+        return MP_SPAWN_CTX
 
         # TODO: how to track helper processes created by subclasses, so that
         # the sys info log in `_gather_output` can include them?
+
+    @classmethod
+    def get_mpcontext(cls):
+        # TODO: remove
+        return cls.get_mp_context()
 
     async def async_call(self, x, /, *, timeout: Union[int, float] = 60):
         # When this is called, it's usually backing a (http or other) service.